requirements.txt
azure-digitaltwins-core
azure-identity
python-dotenv
web3
dash
plotly
matplotlib
ecdsa
traci
sumolib
pandas
